move config tests to target specific subdirectory
